Support for the new Checknum property in transactions
The Checknum property of type decimal was added to the Transaction class to complement the existing functionality. The Checksum property was marked as obsolete with the [Obsolete] attribute.

In the Parser.cs file, the assignment to the Checksum property was corrected using long.TryParse for validation. Support was also added to assign the value to the new Checknum property using decimal.TryParse.